Using 3D points obtained with solvePnP in reference camera (camera 0), instead of multiview triangulation (which is not working well so far).
Please note this is not the best approach because there will be a bias error in the cameras (reprojection error will be very good in the first camera but not so good in the remaining cameras).

Anyway errors in prosicilia camera are too big, probably because it is not well calibrated (intrinsically).